cxy2020年4月30日14点50分

+++ src/main/java/com/ming/hospital/service/DoctorService.java
@@ -27,7 +27,7 @@ public interface DoctorService {
 
     Integer addDoctor(Doctor doctor);
 
-    Boolean del(Integer id);
+    Boolean del(Long id);
 
     Boolean edit(Doctor doctor);
 }
